+++ packages/sphinx-test-reports/sphinxcontrib/test_reports/directives/test_report.py
@@ -4,7 +4,8 @@ import os
 from docutils import nodes
 from docutils.parsers.rst import directives
 
-from sphinxcontrib.test_reports.directives.test_common import TestCommonDirective
+from sphinxcontrib.test_reports.directives.test_common import \
+    TestCommonDirective
 from sphinxcontrib.test_reports.exceptions import InvalidConfigurationError
 
 # fmt: on