+++ packages/sphinx-test-reports/docs/changelog.rst
@@ -7,9 +7,12 @@ Changelog
 -----
 :Released: under development
 
+* Added testcases for the :ref:`json_parser` and updated docs.
 * Updated project documentation to use the Useblocks theme customization.
-* Improvement: Template file encoding coud be configured. See :ref:`tr_import_encoding`.
+* Improvement: Template file encoding could be configured. See :ref:`tr_import_encoding`.
   `#60 <https://github.com/useblocks/sphinx-test-reports/issues/60>`_
+*  Improvement: Supporting JSON files containing test results: :ref:`json_parser`.
+*  Improvement: Implemented :ref:`tr_json_mapping` config option for JSON mapping.
 
 1.0.2
 -----